+++ tests/nghttp2_test_helper.h
@@ -82,12 +82,13 @@ ssize_t inflate_hd(nghttp2_hd_inflater *inflater, nva_out *out,
                    nghttp2_bufs *bufs, size_t offset, nghttp2_mem *mem);
 
 int pack_headers(nghttp2_bufs *bufs, nghttp2_hd_deflater *deflater,
-                 int32_t stream_id, int flags, const nghttp2_nv *nva,
+                 int32_t stream_id, uint8_t flags, const nghttp2_nv *nva,
                  size_t nvlen, nghttp2_mem *mem);
 
 int pack_push_promise(nghttp2_bufs *bufs, nghttp2_hd_deflater *deflater,
-                      int32_t stream_id, int flags, int32_t promised_stream_id,
-                      const nghttp2_nv *nva, size_t nvlen, nghttp2_mem *mem);
+                      int32_t stream_id, uint8_t flags,
+                      int32_t promised_stream_id, const nghttp2_nv *nva,
+                      size_t nvlen, nghttp2_mem *mem);
 
 int frame_pack_bufs_init(nghttp2_bufs *bufs);
